Projects/Batch-2022-2026/Spam-Detection-for-Smart-Home-Devices-using-Machine-Learning/requirements.txt
numpy>=2.0
pandas>=2.0.3
scikit-learn>=1.3.2
flask>=3.0.3
matplotlib
flask-sqlalchemy>=3.0
flask-migrate>=4.0
flask-jwt-extended>=4.6
flask-cors>=4.0
python-dotenv>=1.0
werkzeug
# TensorFlow/Keras - install below if needed
# tensorflow>=2.13.0
# keras>=2.13.1
